// src/shared/api/types/Measurement.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type Measurement = {
	/**
	 * @description The code of the measurement being recorded.
	 * @type string
	 */
	measurementCode: string
	/**
	 * @description The preferred unit of measure for this exercise / measurement.
	 * @type string
	 */
	preferredUnit?: string | null
	/**
	 * @description When used in a WorkoutLog, this is the placeholder value from the workout template.
	 */
	measurementPlaceholder?: (number | string) | null
	measurementValue?: (number | string) | null
}